Move EnumToAvailableValues to a separate file
The descriptions.h file does not get included transitively in main.cpp
anymore.
