+++ src/algorithms/options/enum_to_available_values.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include <sstream>
+#include <string>
+
+namespace algos {
+
+template <typename BetterEnumType>
+static std::string EnumToAvailableValues() {
+    std::stringstream avail_values;
+
+    avail_values << '[';
+
+    for (auto const& name : BetterEnumType::_names()) {
+        avail_values << name << '|';
+    }
+
+    avail_values.seekp(-1, std::stringstream::cur);
+    avail_values << ']';
+
+    return avail_values.str();
+}
+
+}  // namespace algos
